requirements.txt
flask==3.1.0
praw==7.8.1
openai==1.68.2
python-dotenv==1.0.1
pydantic==2.10.6
